task-manager-backend/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-Pydantic==0.13.1
Flask-SQLAlchemy==3.1.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.0
Werkzeug==3.1.3
